crime-heatmap-flask/requirements.txt
Flask-Cors>=4.0.0,<5.0.0
Flask>=3.0.0,<4.0.0
Flask-Migrate>=4.0.7,<5.0.0
Flask-SocketIO>=5.3.6,<6.0.0
Flask-SQLAlchemy>=3.1.1,<4.0.0
Marshmallow>=3.21.1,<4.0.0
Pillow>=9.5.0
python-engineio>=4.9.0,<5.0.0
python-socketio>=5.11.0,<6.0.0
python-dotenv>=1.0.1,<2.0.0
requests>=2.31.0,<3.0.0
SQLAlchemy>=2.0.36,<3.0.0
